ur-licence-mann/tests/requirements.txt

paho-mqtt>=1.6.1
